fix(schema): map stock_sample_id as Integer to match int4 DB column
Hibernate schema validation failed because stock_sample_id was mapped
as Long (→ bigint) while the actual v_allstock_basicprop column is int4.
Changed Germplasm.stockSampleId, GermplasmRepository key type, and all
callSetDbId parse sites to Integer/parseInt.

+++ src/main/java/org/irri/snpseek/brapi/controller/CallSetController.java
@@ -126,9 +126,9 @@ public class CallSetController {
     public BrapiResponse<CallSetDto> getCallSet(
             @Parameter(description = "Numeric callSetDbId (stock_sample_id)", required = true)
             @PathVariable String callSetDbId) {
-        long stockSampleId;
+        int stockSampleId;
         try {
-            stockSampleId = Long.parseLong(callSetDbId);
+            stockSampleId = Integer.parseInt(callSetDbId);
         } catch (NumberFormatException e) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND,
                     "CallSet not found for callSetDbId: " + callSetDbId);
@@ -185,7 +185,7 @@ public class CallSetController {
             }
             if (callSetDbId != null && !callSetDbId.isBlank()) {
                 try {
-                    predicates.add(cb.equal(root.get("stockSampleId"), Long.parseLong(callSetDbId)));
+                    predicates.add(cb.equal(root.get("stockSampleId"), Integer.parseInt(callSetDbId)));
                 } catch (NumberFormatException ignored) {
                     predicates.add(cb.disjunction()); // no match
                 }
